requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.0
charset-normalizer==3.3.2
cryptography==43.0.0
Django==5.1
django-allauth==64.2.0
django-debug-toolbar==4.4.6
django-extensions==3.2.3
django-tinymce==4.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.8
pillow==10.4.0
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.9.0
gunicorn==22.0.0
python-dotenv==1.0.1
python-slugify==8.0.4
requests==2.32.3
sqlparse==0.5.1
text-unidecode==1.3
urllib3==2.2.2
